+++ airbyte-cdk/python/airbyte_cdk/sources/file_based/config/jsonl_format.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing_extensions import Literal
+
+
+class JsonlFormat(BaseModel):
+    class ParquetFormat(BaseModel):
+        filetype: Literal["jsonl"] = "jsonl"
